src/pi1/requirements.txt
smbus2
paho-mqtt
sparkfun-qwiic-mux
pyyaml
boto3
AWSIoTPythonSDK 
